ci(updater): move key expiration canary test to dedicated workflow
Move TestGPGVerifyIn6Months from verify_test.go to verify_canary_test.go
behind the canary build tag. This prevents upcoming key expiration from
failing the standard test suite and obscuring other test results.

Add a dedicated GitHub Actions workflow (key-canary.yml) that runs the
canary test on push, pull request, and a weekly schedule. When the
updater signing key is within 6 months of expiration, the canary workflow
will fail in isolation, providing a clear indication in the Actions
summary.

Closes #3523
